refactor: CustomHttpException을 핸들링하도록 변경
